+++ crates/infrastructure/src/sql/cache.rs
@@ -864,6 +864,10 @@ async fn drain_buffer(pool: &PgPool, buffer: &mut VecDeque<DatabaseQuery>) {
                     DatabaseQueries::add_instrument(pool, "CRYPTO_FUTURE", Box::new(instrument))
                         .await
                 }
+                InstrumentAny::CryptoOption(instrument) => {
+                    DatabaseQueries::add_instrument(pool, "CRYPTO_OPTION", Box::new(instrument))
+                        .await
+                }
                 InstrumentAny::CryptoPerpetual(instrument) => {
                     DatabaseQueries::add_instrument(pool, "CRYPTO_PERPETUAL", Box::new(instrument))
                         .await