+++ nautilus_trader/model/instruments/__init__.py
@@ -24,6 +24,7 @@ from nautilus_trader.model.instruments.binary_option import BinaryOption
 from nautilus_trader.model.instruments.cfd import Cfd
 from nautilus_trader.model.instruments.commodity import Commodity
 from nautilus_trader.model.instruments.crypto_future import CryptoFuture
+from nautilus_trader.model.instruments.crypto_option import CryptoOption
 from nautilus_trader.model.instruments.crypto_perpetual import CryptoPerpetual
 from nautilus_trader.model.instruments.currency_pair import CurrencyPair
 from nautilus_trader.model.instruments.equity import Equity
@@ -41,6 +42,7 @@ __all__ = [
     "Cfd",
     "Commodity",
     "CryptoFuture",
+    "CryptoOption",
     "CryptoPerpetual",
     "CurrencyPair",
     "Equity",