+++ RELEASES.md
@@ -3,6 +3,7 @@
 Released on TBD (UTC).
 
 ### Enhancements
+- Added `CryptoOption` instrument, supporting inverse and fractional sizes
 - Added catalog consolidation functions of several parquet files into one (#2421), thanks @faysou
 - Added `Cache.prices(...)` to return a map of latest price per instrument for a price type
 - Added FDUSD (First Digital USD) crypto `Currency` constant